backend/requirements.txt
fastapi>=0.115.0
python-multipart>=0.0.9
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
google-genai>=1.0.0
slowapi>=0.1.9
pyjwt[crypto]>=2.8.0
pypdf>=5.0.0
pdf2image>=1.17.0
pytesseract>=0.3.13
Pillow>=10.0.0
